Add more tree bindings
Add git_commit_tree() and friend.

Add const variations of several member functions in php_resource_ref and
git2_resource<T> to allow for const handle types being assigned. We
still have to cast away const instead of creating a new resource type to
handle the const version. If we did create a new resource type, then the
resource wouldn't be compatible with any of the functions that require
the non-const version. This is safe since we assign a const resource
using the nofree variation that prevents the user from calling functions
they shouldn't.

Fix problem in test base with showing exception message.