Add php_resource_owned for forcing handle ownership
Add php_resource_owned type to handle cases where we want to only use a
handle that is owned by the resource structure. The use case for this is
very tiny but still relevant.

Fix problem in return value position in git_commit_tree_id() generator.
It was returning the first pack element, not the wrapped function return
value.