Add resource dependency functionality
Add initial resource dependency functionality to the extension. This
tracks resources depended on by other resources, ensuring that the
underlying memory structures are only freed after the dependent
structures have been freed. This ensures that users cannot bork the PHP
process by causing an AV/segfault when using a resource that depends on
a freed resource that's passed out of scope. This is an essential
feature that ensures the production-worthiness of the extension.

Allow each resource to depend on (at most) one other resource. In most
cases, a resource depends on the git_repository that created it, or some
other opaque git2 type. We are super paranoid in setting dependencies,
so we assume dependencies exist whenever objects are created using a
function that requires a git2 handle. We implement freeing the resource
backing objects using dynamic dispatch, which allows us to free parents
recursively of any type.

The 'setdeps' php function generators were added to handle the redundant
operation of setting the dependency chain. It invokes an implementation
function that forwards the required variables to a function that calls
get_object() to retrieve the wrapped resource object and set_parent() to
set the parent resource. Typically this is used with git2_resource<T>
instantiations but can work with any arbitrary type.