Update free() bindings to not make direct calls
The free() bindings shouldn't call the underlying free() git2 functions
directly anymore. Because we now track resource dependencies, we can't
have the user free a handle that is depended on by another resource.

The free() bindings now only invalidate the specified resource. This
process may or may not invalidate the underlying handle.