Set linkage of binding functions to static
This decreases the size of the extension when compiled as a shared
library by allowing the long, mangled C++ template names to be stripped
out. Before this change, these names were locked in the dsym table as
weak symbols and could not be stripped out, thus contributing to a large
executable.