Fix bug where custom resource object dstors not called
This patch fixes a critical bug in the custom resource objects used to
implement async callbacks. The callback object wasn't getting freed
since the compiler was not calling the derived class destructor. This
was fixed by turning ~git2_resource() into a virtual destructor.

Also fixed was the call to decrement the refcounts for the callback's
two zval members. The previous implementation was just decrementing the
ref counts. This patch uses zval_ptr_dtor() to ensure those zvals get
cleaned up.