Add resource owner dependencies to GitODBBackend
This patch allows the GitODBBackend to extend the lifetime of its parent
git_odb instance by tracking the owner parent. This prevents crashes
when users unset the git_odb resource zval but keep a GitODBBackend
instance that accesses a git_odb_backend is managed by the git_odb.

The GitODBBackend object's "odb" property was updated to use the same
git_odb resource wrapper to allow zvals created from this property to
track the original instance (to prevent crashes when the original is
deleted but a zval still references a pointer to freed memory).

The connector_wrapper components were updated to allow multiple
connector wrappers back-to-back. This is performed using
connector_wrapper_ex.