Improve array handling and build system
This patch refactors array handling to its own compilation unit. The
same C-style macros are still in use (with a small change to the API)
but everything is now backed by C++ array_wrapper object that better
encapsulates the functionality.

The build system was also improved. The forcing of CXXFLAGS was taken
out and just the forcing of C++11 mode was retained. It is now up to the
user to specify their own CXXFLAGS at configure-time.