odb-backend: Fix circular reference in ODB owner
This patch fixes a nasty, circular reference in the ODB backend's ODB
owner reference. Since the ODB owns the ODB backend, there is no need to
establish a reference to the ODB from the ODB backend.

Add call to userspace free() method in free handler. This allows the
custom ODB implementation to cleanup resources if needed.