Fix GitClosure get_closure handling
This patch correctly implements get_closure for class GitClosure. The
functionality was updated to set a scope for the wrapped function when
it assigns a 'this' object. Previously this was causing references to
not be decremented. We had fixed this in the last commit erroneously by
using Z_DELREF_P() but this patch actually fixes the root problem.