Separate GitODBBackend into two classes
This patch separates the GitODBBackend functionality into two classes.
One is the base class that is now abstract and the other is
"GitODBBackend_Internal" which is used for internal ODBs. This design
greatly simplifies the code so that we don't have to account for two
different use cases within the implementation: we only have to focus on
one.