Fix bugs in read_property object handlers
Fixes issue causing segfault/av when accessing custom object properties
via read_property handler. This occurred when accessing properties that
employed the object property storage HashTable to cache the property
zval. We forgot to increment the zval's refcount so they table entry was
corrupted when the zval was freed externally.