php-array: Fix handling of function/payload zvals
This patch adds properly sets the function and payload zvals onto a
php_callback_base instance. We added set_members() to php_callback_base
to better abstract what needs to happen. The issue was caused by the
callback instance assuming it had references it hadn't incremented. The
fix ensures the increments occurred when the zvals are set on the
callback instance.