Handle fatal errors from callbacks that git2 doesn't check
There are some callbacks called by libgit2 that are not supposed to
report errors. So when we allow a PHP userspace callback to be used, we
have to have a way to report fatal errors; otherwise, the PHP instance
may crash since its execution state thinks it should be handling a fatal
error. Note that exceptions propagate naturally and do not affect the
PHP engine so we don't have to worry about these.

What we do for fatal errors (i.e. those caught by our
php_bailout_context instance during the callback), is set a global flag
that alerts any top-level php_bailer that a fatal error should be
propagated. This allows the libgit2 call stack to unwind naturally
before the PHP internal function implementation catches the fatal error
via the php_bailer and issues a longjmp to terminate the request
execution.