Ensure owner is set on custom backend objects
This patch ensures that the owner is set on a custom backend when it is
added to its parent. Previously we did not account for the base where
the backend would already exist and so we overlooked setting the owner.
This patch ensures the owner is set in all cases.