Improve local_pack connector abstraction
Improves the local_pack connector mechanism so that the connected type
is extracted directly by the local_pack. This allows connector_wrapper
instances to automatically extract the underlying value from an adjacent
connector_wrapper without need of connector_wrapper_ex.

This also adds an Offset template parameter to the connector_wrapper,
which allows the wrapper to connect a value at an arbitrary position (in
a single direction) in the pack.