Add GIT_EPHP error flag to handle most runtime errors
This error code is to be used by the extension whenever something goes
wrong inside libgit2 code provided by this PHP extension. This updates
the error behavior so that an exception is thrown in userspace instead
of a fatal error.

We will now only use GIT_EPHPFATAL for errors that are unexpected.