Fix alloc/free of custom object types and copy of zval
Fixes a bug causing memory manager corruption. The bug was due to
freeing object memory in the free function that should have been left
to the engine. The free function is only for freeing custom data: PHP
takes care of freeing the top-level structure.

Also fixed some zval copy usage to ensure an accurate copy of the zval
is made.