Move libgit2 initialization/shutdown into request lifecycle
This commit moves libgit2 global initialization and shutdown from the
module lifecycle into the request lifecycle. While this does mean the
extension performs move allocs/frees per request, it allows us to
release all memory back to PHP when the library is not in use.

We made this decision so that it would be easier to implement custom
allocators later on. (Custom allocation becomes possible with libgit2
v0.27.1.) If libgit2 were using PHP's per-request allocator, then any
accidental leaks from the library could be easily cleaned up for a
long-lived PHP process.