Fix bug in testbed repository loading
Since the testbed wipes the disk between each test suite, it's important
that any in-memory structures are not reused between test suites. There
was a bug in the RepositoryTestCase (and similar) that caused an
existing git_repository to stay loaded after tear down since the
variable was declared as a static local. The fix is to move the variable
as a static class property and reset when setting up.

Also, this error was only happening in php 8.1 and not php 8.0, which is
very perplexing. I suspect that the way local static variables are
cleaned up (or perhaps resources) was changed somehow in PHP 8.1, which
exposed this bug.